기존 하드코등으로 컴포넌트 별로 별도의 코드를 가지도 있던 UI 부분을 각 타입 별로 나누어 출력하도록 변경함 -> 어떤 컴포넌트가 와도 문제 없도록 하기 위함

+++ src/ThruthGameEngine/RigidBody.cpp
@@ -88,7 +88,7 @@ void Truth::RigidBody::CalculateMassCenter()
 	{
 		pos += c.lock()->m_center;
 	}
-	pos /= m_colliders.size();
+	pos /= static_cast<float>(m_colliders.size());
 
 	m_body->setCMassLocalPose(
 		physx::PxTransform(MathConverter::Convert(pos)));
